+++ src/main/kotlin/io/github/drewlakee/plugins/json/openapi/PluginIcons.kt
@@ -1,0 +1,8 @@
+package io.github.drewlakee.plugins.json.openapi
+
+import com.intellij.openapi.util.IconLoader
+
+object PluginIcons {
+    @JvmField
+    val toolbarWindow = IconLoader.getIcon("/icons/swagger-icon.svg", javaClass)
+}