+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 <!-- Keep a Changelog guide -> https://keepachangelog.com -->
 
-# idea-jooas-plugin Changelog
+# Changelog
 
-## [Unreleased]
+## [1.0.0] - 2024-09-15
+
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+
+- UI Tool Window for visual conversion 
+- Functionality for converting Json Object into OpenAPI YAML Schema
+- Optional feature toggles for the functionality